Engine prep and clean-up
- remove mouse
- add pawn and move functionality from robocat
- remove collision handling from robocat
- remove health from robocat
- remove code that we won't need
- fix IDE warnings

+++ RoboCatSFML/MoveList.hpp
@@ -20,7 +20,7 @@ public:
 
 	void			Clear() { mMoves.clear(); }
 	bool			HasMoves()				const { return !mMoves.empty(); }
-	int				GetMoveCount()			const { return mMoves.size(); }
+	int				GetMoveCount()			const { return (int)mMoves.size(); }
 
 	//for for each, we have to match stl calling convention
 	const_iterator	begin()					const { return mMoves.begin(); }
