+++ RoboCatSFML/ScoreBoardManager.cpp
@@ -83,7 +83,7 @@ void ScoreBoardManager::IncScore(uint32_t inPlayerId, int inAmount)
 
 bool ScoreBoardManager::Write(OutputMemoryBitStream& inOutputStream) const
 {
-	int entryCount = mEntries.size();
+	int entryCount = (int)mEntries.size();
 
 	//we don't know our player names, so it's hard to check for remaining space in the packet...
 	//not really a concern now though
