+++ RoboCatSFML/World.cpp
@@ -17,7 +17,7 @@ World::World()
 void World::AddGameObject(GameObjectPtr inGameObject)
 {
 	mGameObjects.push_back(inGameObject);
-	inGameObject->SetIndexInWorld(mGameObjects.size() - 1);
+	inGameObject->SetIndexInWorld((int)mGameObjects.size() - 1);
 }
 
 
@@ -25,7 +25,7 @@ void World::RemoveGameObject(GameObjectPtr inGameObject)
 {
 	int index = inGameObject->GetIndexInWorld();
 
-	int lastIndex = mGameObjects.size() - 1;
+	int lastIndex = (int)mGameObjects.size() - 1;
 	if (index != lastIndex)
 	{
 		mGameObjects[index] = mGameObjects[lastIndex];
@@ -42,7 +42,7 @@ void World::Update()
 {
 	//update all game objects- sometimes they want to die, so we need to tread carefully...
 
-	for (int i = 0, c = mGameObjects.size(); i < c; ++i)
+	for (int i = 0, c = (int)mGameObjects.size(); i < c; ++i)
 	{
 		GameObjectPtr go = mGameObjects[i];
